Move actuals-adding into its own class
