+++ Pipeline.py
@@ -19,19 +19,6 @@ class Pipeline():
         self.terminals_path = terminals_path
         self.vehicles_path = vehicles_path
 
-    # Build a dataframe from prediction analyzer logs, dropping actuals without
-    # times, commuter rail trips, and duplicates (which we get because we log
-    # both from dev-green and prod).
-    def load_actuals(self):
-        raw_frame = pd.read_csv(
-            self.actuals_path,
-            usecols=['event_type', 'stop_id', 'time', 'trip_id', 'vehicle_id'],
-            dtype={"stop_id": str}
-        )
-        dropped_frame = raw_frame.dropna(subset=["time"])
-        is_subway = dropped_frame["trip_id"].apply(lambda x: x[0:3] != "CR-")
-        return dropped_frame[is_subway].drop_duplicates()
-
     # Build a dictionary mapping location IDs to GTFS stop IDs, using parent
     # stops where child stops exist.
     def location_id_to_stop_id(self):
@@ -58,41 +45,11 @@ class Pipeline():
     # stop GTFS ID
     def gtfs_id_for_terminals(self):
         patterns = self.load_patterns()
-        locations = self.location_id_to_stop_id()
+        locations = self.location_id_to_parent_stop_id()
         patterns["terminal_gtfs_id"] = patterns["terminal_stop"] \
             .apply(lambda loc_id: locations.get(loc_id))
         patterns = patterns.dropna().drop("terminal_stop", axis=1)
         return patterns
-
-    # Build a dataframe with every combination of given vehicle datapoints
-    # and possible destination.
-    #
-    # Note that we don't currently check that the destination makes sense, for
-    # example, there will be rows for the Blue Line with a destination of
-    # Kendall. Currently these get eliminated in a later step. It might be
-    # more efficient to do so up-front here, though we'd then have to have this
-    # code know somehow what stops are on what lines.
-    def add_all_possible_destinations(self):
-        vehicle_datapoints = self.load_vehicle_datapoints()
-        gtfs_ids = set(self.location_id_to_stop_id().values())
-        blank_frame = pd.DataFrame()
-        new_frames = map(
-            lambda gtfs_id: self._vehicle_datapoints_with_destination_gtfs_id(
-                vehicle_datapoints,
-                gtfs_id
-            ),
-            gtfs_ids
-        )
-        return blank_frame.append(list(new_frames))
-
-    def _vehicle_datapoints_with_destination_gtfs_id(
-        self,
-        vehicle_datapoints,
-        gtfs_id
-    ):
-        new_frame = vehicle_datapoints.copy()
-        new_frame["destination_gtfs_id"] = str(gtfs_id)
-        return new_frame
 
     # Build a dataframe with all logged terminal-mode datapoints
     def load_terminal_datapoints(self):
@@ -210,31 +167,6 @@ class Pipeline():
                 generational_n_hot_location_lists[generation] = n_hot_columns
 
         return generational_n_hot_location_lists
-
-    # Join actuals to a set of vehicle datapoints
-    def add_actuals(self, vehicle_datapoints):
-        actuals = self.load_actuals()
-        actuals = actuals.rename(
-            {"trip_id": "gtfs_trip_id", "stop_id": "destination_gtfs_id"},
-            axis="columns"
-        )
-        actuals["gtfs_trip_id"] = actuals["gtfs_trip_id"].astype(str)
-        vehicle_datapoints["gtfs_trip_id"] = \
-            vehicle_datapoints["gtfs_trip_id"].astype(str)
-        return vehicle_datapoints.merge(
-            actuals,
-            how="inner",
-            on=["gtfs_trip_id", "destination_gtfs_id"]
-        ).drop(
-            [
-                "generation",
-                "gtfs_trip_id",
-                "ocs_trip_id",
-                "vehicle_id_x",
-                "vehicle_id_y"
-            ],
-            axis=1
-        )
 
     def _parent_stop(self, acc, x):
         location_id = x[0]
